The "QcMetadata" class

The QcMetadata class is a simple interface to metadata. The objects can be displayed with show for a summary and print for the content.

Objects from the Class

Objects can be created by calls of the form QcMetadata(...).

Slots

metadata:

Object of class "list" that stores the metadata variables. The list must be named. NA and empty characters are not allowed.

Methods

[

signature(x = "QcMetadata"): subsets x as a new QcMetadata instance.

[[

signature(x = "QcMetadata"): extracts a single element of x.

metadata

signature(x = "QcMetadata"): return the object's metadata list. Also available as mdata.

metadata<-

signature(x = "QcMetadata", value = "list"): sets the objects metadata. Also available as mdata.

length

signature(x = "QcMetadata"): returns the number of metadata variables.

names

signature(x = "QcMetadata"): returns the names of the metadata variables.

Examples

QcMetadata(metadata =
           list(name = "John Doe",
                lab = "Big Lab in Big Uni"))

#> Object of class "QcMetadata"
#>  containing 2 variables.

## less typing
qmd <- QcMetadata(list(name = "John Doe",
                       lab = "Big Lab in Big Uni"))
mdata(qmd)

#> $name
#> [1] "John Doe"
#> 
#> $lab
#> [1] "Big Lab in Big Uni"
#> 

show(qmd)

#> Object of class "QcMetadata"
#>  containing 2 variables.

print(qmd)

#> name 
#>  [1] "John Doe"
#> lab 
#>  [1] "Big Lab in Big Uni"