N15 labelling QC report

A simple wrapper for the QC of 15N labelling. The respective QC items are the distribution of PSM incorporation rates, distribution of log2 fold-changes and number of identified features. See the vignette for details.

n15qc(object,
    fcol = c("Protein_Accession", "Peptide_Sequence", "Number_Of_Unique_Peptides", "Variable_Modifications", "inc"),
    inctr = 97.5, lfctr = c(-0.5, 0.5), type, reportname)

Arguments

object
An MSnSet to be quality controlled.

fcol
The name of the feature variables for the protein identifiers (accession numbers for example), the peptide sequences, the number of unique peptides for each identified protein, the variable modifications identified on the peptides and the N15 incorporation rate. These must be provided in that order. Defaults are Protein_Accession, Peptide_Sequence, Number_Of_Unique_Peptides, Variable_Modifications, and inc.

inctr
The minimum level of median incorporation rate to set the QC item status to TRUE. Default is 97.5.

lfctr
The range of accepted median PSM log2 fold-change for the QC item status to be set to TRUE. Default is c(-0.5, 0.5).

type
The type of report to be saved. If missing (default), no report is generated. See qcReport for details.

reportname
The name of the report, in case a type is defined. If missing (default), the report will be names n15qcreport followed by the generation data and time.

Value

Invisibly returns the resulting QcMetrics instance.